structure current code to different modules and fix linters
