+++ modules/initializeBookList.js
@@ -1,4 +1,4 @@
-import BookList from '../index.js';
+import BookList from './bookList.js';
 
 export default function initializeBookList() {
   const bookList = new BookList();
